Move files to own namespaces
